pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-machine"
version = "0.0.1"
description = "Kalshi weather market making"
requires-python = ">=3.10"
dependencies = ["requests", "python-dotenv", "cryptography"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
